# Optimize the locking strategy used in NonceService and stop using math/big