Updates #1219 and adds two benchmarks, could use a serious once over.

`master`

```
$ go test -bench=. -cpu 8 -test.benchmem
PASS
BenchmarkNonce-8	  500000	      2606 ns/op	     283 B/op	      10 allocs/op
BenchmarkValid-8	 1000000	      1020 ns/op	     200 B/op	       6 allocs/op
ok  	github.com/letsencrypt/boulder/nonce	3.918s
```

`e9501b5`

```
$ go test -bench=. -cpu 8 -test.benchmem
PASS
BenchmarkNonce-8	  500000	      2261 ns/op	     174 B/op	       5 allocs/op
BenchmarkValid-8	 3000000	       429 ns/op	     120 B/op	       4 allocs/op
ok  	github.com/letsencrypt/boulder/nonce	3.949s
```