@jsha 2016-06-06
I'm not sure this is justified - we're adding significant complexity to the locking scheme in exchange for saving half a microsecond per request, according to your benchmark. I'm inclined to leave this unchanged unless we see this turn up as a performance bottleneck during load testing.