@ccppuu 2016-06-07
Holding off on reviewing this until there's consensus on whether its going to be a bottleneck worth addressing.