@rolandshoemaker 2016-06-07
I'm going to close this out for now, I'll do some testing while working on the load generators and if this shows up as a big performance gain I'll re-open.